feat: add umbrella chart for the orchestrator
This changei adds a new "umbrella" Helm chart for the Orchestrator
to compile all of its dependencies. Now only Cloudify's chart has
been included as a dependency.

This chart is intended to be installed as a whole to deploy all the
components related to the Orchestrator. Additionally it includes a
"values/" directory aimed to store project-specific values and
configuration.