feat(cloudify): set new admin password from secret
Cloudify's default password is replaced using the scripts mounted into
the pod (unless this behaviour is desabled from values). Up to now, the
new password was configured from chart's values.
Now, the new password will be set from a secret (which has to be created
prior to chart installation).
Chart version is upgraded to relect this change.