feat(keycloak): enable TLS security for Nginx ingress
In the default setup, Nginx comes with self signed TLS certificates.
For live environments it is required to enable Nginx to use production
ready TLS certificates. This is added via cert-manager, which
automatically generates and rotates TLS certificates (it will ensure
certificates are valid and up to date periodically, and attempt to
renew certificates at an appropriate time before expiry) from Let’s
Encrypt.

For that, it is needed to configure an "Issuer" custom resource
(corresponding CRDs must be installed previously) and to add
annotations to the ingress resource so that cert-manager creates the
required certificates for each endpoint.

TLS option can be enabled/disabled from chart values. TLS is disabled by
default as it is intended to be "activated" from the orchestrator chart.